+++ Conf.h
@@ -34,8 +34,8 @@ public:
   std::string  getCallsign() const;
   std::string  getSuffix() const;
   std::string  getRptAddress() const;
-  unsigned int getRptPort() const;
-  unsigned int getMyPort() const;
+  unsigned short getRptPort() const;
+  unsigned short getMyPort() const;
   bool         getDebug() const;
   bool         getDaemon() const;
 
@@ -47,7 +47,7 @@ public:
   bool         getLogFileRotate() const;
 
   // The Network section
-  unsigned int   getNetworkPort() const;
+  unsigned short getNetworkPort() const;
   std::string    getNetworkHosts1() const;
   std::string    getNetworkHosts2() const;
   unsigned int   getNetworkReloadTime() const;
@@ -58,15 +58,15 @@ public:
 
   // The Remote Commands section
   bool           getRemoteCommandsEnabled() const;
-  unsigned int   getRemoteCommandsPort() const;
+  unsigned short getRemoteCommandsPort() const;
 
 private:
   std::string  m_file;
   std::string  m_callsign;
   std::string  m_suffix;
   std::string  m_rptAddress;
-  unsigned int m_rptPort;
-  unsigned int m_myPort;
+  unsigned short m_rptPort;
+  unsigned short m_myPort;
   bool         m_debug;
   bool         m_daemon;
 
@@ -76,7 +76,7 @@ private:
   std::string  m_logFileRoot;
   bool         m_logFileRotate;
 
-  unsigned int   m_networkPort;
+  unsigned short m_networkPort;
   std::string    m_networkHosts1;
   std::string    m_networkHosts2;
   unsigned int   m_networkReloadTime;
@@ -85,8 +85,8 @@ private:
   bool           m_networkRevert;
   bool           m_networkDebug;
 
-  bool         m_remoteCommandsEnabled;
-  unsigned int m_remoteCommandsPort;
+  bool           m_remoteCommandsEnabled;
+  unsigned short m_remoteCommandsPort;
 };
 
 #endif
